+++ lib/elixir/test/elixir/file_test.exs
@@ -189,9 +189,9 @@ defmodule FileTest do
 
         { :ok, files } = File.cp_r(src, dest)
         assert length(files) == 8
-        assert tmp_path("tmp/cp_r/a") in files
-        assert tmp_path("tmp/cp_r/c") in files
-        assert tmp_path("tmp/cp_r/a/1.txt") in files
+        assert tmp_path("tmp/cp_r/a") inlist files
+        assert tmp_path("tmp/cp_r/c") inlist files
+        assert tmp_path("tmp/cp_r/a/1.txt") inlist files
 
         assert { :ok, 'certainly/invalid' } = :file.read_link(tmp_path("tmp/cp_r/c"))
 
@@ -878,8 +878,8 @@ defmodule FileTest do
 
       { :ok, files } = File.rm_rf(fixture)
       assert length(files) == 8
-      assert fixture in files
-      assert tmp_path("tmp/a/1.txt") in files
+      assert fixture inlist files
+      assert tmp_path("tmp/a/1.txt") inlist files
 
       refute File.exists?(tmp_path("tmp/a/1.txt"))
       refute File.exists?(tmp_path("tmp/a/a/2.txt"))
@@ -912,8 +912,8 @@ defmodule FileTest do
 
       { :ok, files } = File.rm_rf(fixture)
       assert length(files) == 8
-      assert fixture in files
-      assert (tmp_path("tmp/a/1.txt") /> to_char_list) in files
+      assert fixture inlist files
+      assert (tmp_path("tmp/a/1.txt") /> to_char_list) inlist files
 
       refute File.exists?(tmp_path("tmp/a/1.txt"))
       refute File.exists?(tmp_path("tmp/a/a/2.txt"))
@@ -948,8 +948,8 @@ defmodule FileTest do
 
       files = File.rm_rf!(fixture)
       assert length(files) == 8
-      assert fixture in files
-      assert tmp_path("tmp/a/1.txt") in files
+      assert fixture inlist files
+      assert tmp_path("tmp/a/1.txt") inlist files
 
       refute File.exists?(tmp_path("tmp/a/1.txt"))
       refute File.exists?(tmp_path("tmp/a/a/2.txt"))